FILE: moves models folder into config-form folder
